TST GH26807 split out test_split_partition
